Peak density parameters

The method creates a template with the parameters for correspondence. Extracted from the function xcms::PeakDensityParam.

PeakDensityPar(...)

Arguments

...
Arguments passed on to xcms::PeakDensityParam

sampleGroups

A vector of the same length than samples defining the sample group assignments (i.e. which samples belong to which sample group). This parameter is mandatory for the PeakDensityParam and has to be provided also if there is no sample grouping in the experiment (in which case all samples should be assigned to the same group).

bw

numeric(1) defining the bandwidth (standard deviation ot the smoothing kernel) to be used. This argument is passed to the [density() method.

minFraction

numeric(1) defining the minimum fraction of samples in at least one sample group in which the peaks have to be present to be considered as a peak group (feature).

minSamples

numeric(1) with the minimum number of samples in at least one sample group in which the peaks have to be detected to be considered a peak group (feature).

binSize

numeric(1) defining the size of the overlapping slices in mz dimension.

maxFeatures

numeric(1) with the maximum number of peak groups to be identified in a single mz slice.

Value

PeakDensityParam S4 object

Slots

.__classVersion__,sampleGroups,bw,minFraction,minSamples,binSize,maxFeatures

See corresponding parameter above. .__classVersion__ stores the version from the class. Slots values should exclusively be accessed via the corresponding getter and setter methods listed above.

References

Colin A. Smith, Elizabeth J. Want, Grace O'Maille, Ruben Abagyan and Gary Siuzdak. "XCMS: Processing Mass Spectrometry Data for Metabolite Profiling Using Nonlinear Peak Alignment, Matching, and Identification" Anal. Chem. 2006, 78:779-787.

See also

Examples

p <- PeakDensityPar(sampleGroups = c(1,2))

#> Warning: mzR has been built against a different Rcpp version (1.0.2)
#> than is installed on your system (1.0.5). This might lead to errors
#> when loading mzR. If you encounter such issues, please send a report,
#> including the output of sessionInfo() to the Bioc support forum at 
#> https://support.bioconductor.org/. For details see also
#> https://github.com/sneumann/mzR/wiki/mzR-Rcpp-compiler-linker-issue.