Chromatographic peak detection (CentWave)

The findChromPeaks_cwp function performs the chromatographic peak detection on LC/GC-MS data. The standard method for peak detection is 'CentWave'. Peak detection aims to detect important features (peaks) on the chromatographic axis. This will be useful for a posterior peak alignment on the chormatophic axis.

find_peaks_cwp(dataset, params)

Arguments

dataset
A lcms_dataset.

params
A converted parameters template from IPO parameters.

Value

A lcms_dataset with the detected peaks added.

Details

Note: signal processing generally consists in three main steps: (1) peak detection (lcms_find_chromp_peaks_cwp function), (2) peak alignment (obiAdjust_Rtime function) and (3) peak correspondence (IPO_group_peaks function). The optimized set of parameters for this signal preprocessing can be obatained from IPO Package.

See also

Examples

if (FALSE) {
file_path <- system.file("extdata", "rearrange_mait", package = "AlpsLCMS")
samples_mzxml <- list.files(file_path, recursive = TRUE, full.names = TRUE)
meta_path <- system.file("extdata", "metadata.xlsx", package = "AlpsLCMS")
opt_result_path <-  system.file("extdata", package = "AlpsLCMS")
preproc_params <- lcms_read_ipo_to_xcms(opt_result_path)

dataset <- suppressWarnings(lcms_read_samples(samples_mzxml, mode = "onDisk"))
metadata <- lcms_meta_read(meta_path)
dataset_meta <- lcms_meta_add(dataset, metadata, by = "sampleNames")

peakdet <- find_peaks_cwp(dataset_meta, params = preproc_params)
print(peakdet)
}