Peak picking optimization

The function optimize parameters considering a set of samples for the peak picking algorithm using the IPO Package. This includes Peak Detection (‘Centwave), Retention Time Correction (‘obiwarp’) and Peak Correspondence (‘Density’).

lcms_peakpicking_optimization(
  dataset,
  peakpickingParameters,
  nSlaves = 1,
  opt_path,
  subdir = "plot_ipo",
  plots = TRUE
)

Arguments

dataset
A lcms_dataset

peakpickingParameters
Parameters for peak picking

nSlaves
Number of slaves the optimization process should spawn.

opt_path
Path where optimization samples are saved.

subdir
Folder where surface plots are saved. If NULL they are displayed by the graphical device.

plots
Defines if plots should be generated (TRUE) or not (FALSE) in a subfolder called "plot_ipo" (default).

Value

A peak picking list with the best setting.

See also

Examples

if (FALSE) {
opt_path <- system.file("extdata", "ipo_opt", package = "AlpsLCMS")
file_name <- system.file("extdata", "dataset_pos_rt_rs.rds", package = "AlpsLCMS")
dataset <- lcms_dataset_load(file_name)
default_peakpicking_params <- lcms_default_peakpicking_params(optimize = TRUE)
result_peakpicking <- lcms_peakpicking_optimization(dataset,
                                                   default_peakpicking_params,
                                                   opt_path = opt_path,
                                                   subdir = NULL)
}