Optimization of retention time correction and grouping parameters.

The function optimizes parameters considering a set of samples for the retention time correction and grouping using the IPO Package.

lcms_retcorgroup_optimization(
  optimizedXcmsSetObject,
  retcorGroupParameters,
  nSlaves = 1,
  opt_path,
  subdir = "plot_ipo",
  plots = TRUE
)

Arguments

optimizedXcmsSetObject
XCMS object conatining the best_settings parameters. This object may be created after running lcms_peakpicking_optimization and extract best_settings$xset (e.g. optimizedXcmsSetObject <- resultPeakpicking$best_settings$xset)

retcorGroupParameters
Parameters for retention time correction and optimization

nSlaves
Number of slaves the optimization process should spawn.

opt_path
Path where optimization samples are saved.

subdir
Folder where surface plots are save. If NULL they are displayed by the graphical device.

plots
Defines if plots should be generated (TRUE) or not (FALSE) in a subfolder called "plot_ipo"(default).

Value

A list with the optimization of parameters for retention time and grouping.

See also

Examples

if (FALSE) {
file_name_pp <- system.file("extdata", "result_peakpicking.rds", package = "AlpsLCMS")
optimizedXcmsSetObject <-base::readRDS(file_name_pp)$best_settings$xset
file_name_rcg <- system.file("extdata", "default_retcorgroup_params.rds", package = "AlpsLCMS")
default_retcorgroup_params <- base::readRDS(file_name_rcg)
opt_path <-  system.file("extdata", package = "AlpsLCMS")

result_retcorgroup <- lcms_retcorgroup_optimization(optimizedXcmsSetObject,
                                                   default_retcorgroup_params,
                                                   opt_path = opt_path,
                                                   subdir = NULL)
}