Feature table

feature_values is a wrapper from xcms. The function creates an intensity matrix with all the features.

feature_values(
object,
method = c("medret", "maxint", "sum"),
value = "into",
intensity = "into",
filled = TRUE,
missing = NA,
msLevel = integer()
)

Value

For featureValues: a matrix with feature values, columns representing samples, rows features. The order of the features matches the order found in the featureDefinitions(object) DataFrame. The rownames of the matrix are the same than those of the featureDefinitions DataFrame. NA is reported for features without corresponding chromatographic peak in the respective sample(s).

See also